requirements.txt
redis==5.2.1
djangorestframework==3.15.2
Markdown==3.7
django-filter==25.1
psycopg2-binary==2.9.10
stripe==11.6.0